z8000: fix sbrk, syscall returns, relocations, and shell for Z8001
Three classes of bugs prevented the V7 shell from running on Z8001:

1. Kernel u_retval was susize_t (16-bit on Z8000) but syscall return
   path does a 32-bit load (ldl rr2). Changed u_retval to arg_t and
   updated all assembly offsets in kernel-z8000.def. Also fixed _sbrk
   overflow check to use arg_t instead of ssize_t.

2. Libc syscall wrappers: on big-endian Z8001, a 32-bit return value
   of 1 is 0x00000001 (r2=0, r3=1), but C functions returning int
   (16-bit) expect the result in r2. Added 'ld r2, r3' to all wrappers
   except pointer-returning syscalls (sbrk, signal, _sigdisp).

3. coff2aout: R_IMM32_NORELAX (0x31) relocations in .data section were
   silently dropped, causing global pointers to have segment 0. Now
   handled identically to R_IMM32.

4. V7 shell Lcheat macro used (int *) cast — on Z8001 where int is
   16-bit, this only accesses the segment word of a 32-bit pointer.
   Setting ARGMK bit corrupted segments. Changed to (intptr_t *).

Also enabled DEBUG section in blok.c and fixed chkbptr for Z8001.

+++ Applications/V7/cmd/sh/mode.h
@@ -48,7 +48,7 @@ typedef struct namnod *NAMPTR;
  * into an Rvalue so two cheats
  * are necessary, one for each context.
  */
-#define Lcheat(a)	(*(int *)&(a))
+#define Lcheat(a)	(*(intptr_t *)&(a))
 
 
 /* address puns for storage allocation */
